Create gcp wif helpers file and remove unnecessary validation functions
