+++ cmd/ocm/gcp/update-wif-config.go
@@ -7,7 +7,6 @@ import (
 
 	"github.com/openshift-online/ocm-cli/pkg/gcp"
 	"github.com/openshift-online/ocm-cli/pkg/ocm"
-	cmv1 "github.com/openshift-online/ocm-sdk-go/clustersmgmt/v1"
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 )
@@ -18,26 +17,21 @@ var UpdateWifConfigOpts struct {
 // NewUpdateWorkloadIdentityConfiguration provides the "gcp update wif-config" subcommand
 func NewUpdateWorkloadIdentityConfiguration() *cobra.Command {
 	updateWifConfigCmd := &cobra.Command{
-		Use:     "wif-config [ID|Name]",
-		Short:   "Update wif-config.",
-		RunE:    updateWorkloadIdentityConfigurationCmd,
-		PreRunE: validationForUpdateWorkloadIdentityConfigurationCmd,
+		Use:   "wif-config [ID|Name]",
+		Short: "Update wif-config.",
+		RunE:  updateWorkloadIdentityConfigurationCmd,
 	}
 
 	return updateWifConfigCmd
 }
 
-func validationForUpdateWorkloadIdentityConfigurationCmd(cmd *cobra.Command, argv []string) error {
-	if err := wifKeyArgCheck(argv); err != nil {
-		return err
-	}
-	return nil
-}
-
 func updateWorkloadIdentityConfigurationCmd(cmd *cobra.Command, argv []string) error {
 	ctx := context.Background()
 	log := log.Default()
-	key := argv[0]
+	key, err := wifKeyFromArgs(argv)
+	if err != nil {
+		return err
+	}
 
 	// Create the client for the OCM API:
 	connection, err := ocm.NewConnection().Build()
@@ -80,27 +74,4 @@ func updateWorkloadIdentityConfigurationCmd(cmd *cobra.Command, argv []string) e
 	}
 
 	return nil
-}
-
-// findWifConfig finds the WIF configuration by ID or name
-func findWifConfig(client *cmv1.Client, key string) (*cmv1.WifConfig, error) {
-	collection := client.GCP().WifConfigs()
-	page := 1
-	size := 1
-	query := fmt.Sprintf(
-		"id = '%s' or display_name = '%s'",
-		key, key,
-	)
-
-	response, err := collection.List().Search(query).Page(page).Size(size).Send()
-	if err != nil {
-		return nil, err
-	}
-	if response.Total() == 0 {
-		return nil, fmt.Errorf("WIF configuration with identifier or name '%s' not found", key)
-	}
-	if response.Total() > 1 {
-		return nil, fmt.Errorf("there are %d WIF configurations found with identifier or name '%s'", response.Total(), key)
-	}
-	return response.Items().Slice()[0], nil
 }
